+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1808,7 +1808,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KRGC9R7QKXE7H7B3YVCJ56M7"
+    id: Optional[str] = "01KRGKTFD2JK7M2Y1CVAETFRSN"
 
     description: Optional[str] = None