+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRGC9RQSW8RWJXYDM6MZ0398"
+        "01KRGKTFXCJTWCYTYEA9Y7ZED9"
     )
     r"""The unique identifier of the data source"""