+++ packages/orq-rc/RELEASES.md
@@ -7879,3 +7879,13 @@ Based on:
 - [python v4.10.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.1 - packages/orq-rc
+
+## 2026-05-13 12:12:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.2 - packages/orq-rc